+++ abstractions/ruby/microsoft_kiota_abstractions/lib/microsoft_kiota_abstractions/request_info.rb
@@ -28,11 +28,15 @@ module MicrosoftKiotaAbstractions
       @headers[@@content_type_header] = @@binary_content_type
     end
 
-    def set_content_from_parsable(value, serializer_factory, content_type)
+    def set_content_from_parsable(serializer_factory, content_type, values)
       begin
         writer  = serializer_factory.get_serialization_writer(content_type)
         headers[@@content_type_header] = content_type
-        writer.write_object_value(nil, value);
+        if values != nil && values.kind_of?(Array)
+          writer.write_collection_of_object_values(nil, values)
+        else
+          writer.write_object_value(nil, values);
+        end
         this.content = writer.get_serialized_content();
       rescue => exception
         raise Exception.new "could not serialize payload"